+++ ChoGGi's Library/Code/Classes_UI.lua
@@ -714,6 +714,7 @@ function ChoGGi_XWindow:AddCloseXButton(params)
 
 			-- stuff from my other dialogs
 			if self.CloseXButtonFunc then
+				-- Close other dialogs (from Examine)
 				abort = self:CloseXButtonFunc(...)
 			end