+++ Mods ChoGGi/Expedition Rocket Launch Button/metadata.lua
@@ -1,22 +1,23 @@
 return PlaceObj("ModDef", {
 	"title", "Expedition Rocket Launch Button",
-	"version", 1,
-	"version_major", 0,
-	"version_minor", 1,
-
-	"image", "Preview.png",
 	"id", "ChoGGi_ExpeditionRocketLaunchButton",
 	"steam_id", "1776123839",
 	"pops_any_uuid", "1d8fc47a-1956-462e-83f1-adc1b353916a",
-	"author", "ChoGGi",
 	"lua_revision", 1007000, -- Picard
+	"version", 1,
+	"version_major", 0,
+	"version_minor", 1,
+	"image", "Preview.jpg",
+	"author", "ChoGGi",
 	"code", {
 		"Code/Script.lua",
 	},
 --~ 	"has_options", true,
 	"TagInterface", true,
-	"description", [[Now you can skip unloading cargo you don't care about (and maybe loading?).
+	"description", [[
+Now you can skip unloading cargo you don't care about (and maybe loading?).
 
 
-Requested by BrowncoatTrekky.]],
+Requested by BrowncoatTrekky.
+]],
 })